refactor(config): finish the StartupScriptProvider migration to common
StartupScriptProvider existed twice, in dazzleduck-sql-common and
dazzleduck-sql-flight, and the migration between them was half done: the flight
ConfigBasedStartupScriptProvider was already @Deprecated and delegating to
common's, but the INTERFACE was not, and three of the four call sites still
imported flight's.

common is the right home, as that deprecation note said. Reading SQL to run at
connection setup has nothing to do with Arrow Flight, and a service needing only
that had to depend on the whole Flight module to get it.

flight.StartupScriptProvider is now a @Deprecated(forRemoval) shim extending
common's, keeping its constant and static helpers as delegates so an existing
`implements io.dazzleduck.sql.flight.StartupScriptProvider` still compiles and
`class = "io.dazzleduck.sql.flight.ConfigBasedStartupScriptProvider"` still
loads. Callers in the compactor, the flight server and the runtime now import
common's.

Two call sites were bypassing the interface's own load() and calling
ConfigBasedProvider.load(config, PREFIX, new ConfigBasedStartupScriptProvider())
— re-deriving, at the call site, exactly the arguments StartupScriptProvider.load
already supplies. They now call load(config). That also removes the last thing
tying them to ConfigBasedProvider, which common cannot extend anyway: common
sits BELOW commons, so the duplicated reflection there is structural, not an
oversight.

Consequence worth the change on its own: the compactor's ONLY use of Flight was
this interface, so it no longer depends on dazzleduck-sql-flight at all. A
compaction daemon no longer drags in an Arrow Flight server. It declares
dazzleduck-sql-common explicitly alongside commons, since it imports from both
directly and a transitive path is not a contract.

Verified on the four touched modules: 0 failures. Two unrelated suites fail
identically on a clean upstream/main in full-reactor scope
(LogForwardingIntegrationTest, and a temp-dir cleanup flake) and are untouched
by this change.
